-demo mode if no dongle -map display -extract axxis orientation from geometry -don't unselect points in crossplotviewer when received point has either inline or crossline not specified

+++ SeismicUI/gatherlabel.cpp
@@ -83,7 +83,7 @@ void GatherLabel::setHighlightTrace( bool on){
     updateBuffers();
 }
 
-void GatherLabel::setHighlightedTrace( size_t t){
+void GatherLabel::setHighlightedTrace( int t){
 
     if( t == m_highlightedTrace) return;
